requirements_mlmodel.txt
flask==3.0.0
pandas==2.1.3
joblib==1.3.2
numpy
scikit-surprise